Fix bug
In a previous commit, we made a minor change in the model which introduced a bug in generation. There, we substituted a transpose in the forward of the encoder by tbc=False in the forward of the model. The forward of the model is not called during generation, so a dimensionality issue appeared. In this commit, we move tbc=False to the forward of the encoder.